requirements.txt
tensorflow>=2.0.0
pandas>=1.0.0
numpy>=1.18.0
matplotlib>=3.1.0
scikit-learn>=0.22.0
pandas-datareader>=0.9.0
python-dotenv>=0.15.0
